llama-project/requirements.txt
Flask
llama-cpp-python
huggingface-hub
